remove parent interface `Iterable` from `ConditionEvents`
This is an overly broad interface for our use case. In particular since we also have `getViolating()` which is the set we are always ultimately interested in. The allowed events are more of a current implementation detail that we need for the inversion logic, but the goal should be to remove this from the public interface eventually.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>